APIs for the instruments in the frequency multicplication spectrometer
Last version (hopefully) of every API needed, containing the API for :
- the synthesizer
- the lock-in amplifier
- the pressure gauge
- the (future) gauge controller
some changes might be needed, especially for the "presets" dictionary where the parameters for the data transmission has been defined and might not correspond to the actual parameters (it should be good for our use though). Concerning the commands, everything should be good too, some might not be needed for a first test or not needed at all, depending on what we can and want to do.