Pull Request for the APIs (#4)
* Create API_MAP_

* Rename API_MAP_ to API_MAP_Rohde_Schwarz_SMF100A

* Update API_MAP_Rohde_Schwarz_SMF100A

* Update API_MAP_Rohde_Schwarz_SMF100A

* new API and change in the synAPI class

- Creation of API for R&S_SMF100A and Pfeiffer_EVR116
- new functions (set/get_lf_freq/waveform) in synthesizer, sinAPI class
- new "blank" commands in the Agilent_E8257D API to fit the new functions

* Update API_MAP_Rohde_Schwarz_SMF100A.yaml

just erased some irrelevant comments

* pfeiffer's gauge API added

needs to be discussed later

* Create API for the MKS gauge

* lock-in API added + minor changes

- lock-in API added
- minor changes, such as dtypes and format command in every files

* Update API_MAP_Ametek_7230.yaml

* APIs for the instruments in the frequency multicplication spectrometer

Last version (hopefully) of every API needed, containing the API for :
- the synthesizer
- the lock-in amplifier
- the pressure gauge
- the (future) gauge controller
some changes might be needed, especially for the "presets" dictionary where the parameters for the data transmission has been defined and might not correspond to the actual parameters (it should be good for our use though). Concerning the commands, everything should be good too, some might not be needed for a first test or not needed at all, depending on what we can and want to do.